Make interface names more consistent.
