+++ src/VersionParser.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * VersionParser interface
+ * Version parser.
  *
  * @package SatisPress
  * @license GPL-2.0-or-later
@@ -12,12 +12,30 @@ declare ( strict_types = 1 );
 namespace SatisPress;
 
 /**
- * Version parser interface.
+ * Version parser class.
  *
  * @package SatisPress
  * @since 0.3.0
  */
-interface VersionParser {
+final class VersionParser implements VersionParserInterface {
+	/**
+	 * Version parser instance.
+	 *
+	 * @var \Composer\Semver\VersionParser
+	 */
+	protected $parser;
+
+	/**
+	 * Initialize the version parser.
+	 *
+	 * @since 0.3.0
+	 *
+	 * @param \Composer\Semver\VersionParser $parser Version parser.
+	 */
+	public function __construct( \Composer\Semver\VersionParser $parser ) {
+		$this->parser = $parser;
+	}
+
 	/**
 	 * Normalizes a version string to be able to perform comparisons on it.
 	 *
@@ -29,5 +47,7 @@ interface VersionParser {
 	 * @param string $full_version Optional complete version string to give more context.
 	 * @return string Normalized version string.
 	 */
-	public function normalize( string $version, string $full_version = null );
+	public function normalize( string $version, string $full_version = null ): string {
+		return $this->parser->normalize( $version, $full_version );
+	}
 }
